Retype Property methods

+++ Property.h
@@ -5,7 +5,7 @@ protected:
 	unsigned int worth;
 public:
 	Property(unsigned int worth);
-	virtual int ñalculationTax() = 0;
-	int calculatIonIncomeTax();
+	virtual double ñalculationTax() = 0;
+	double calculatIonIncomeTax();
 };
